pre-processing ingredients dataset and updated DB with skincare routine
