+++ backend/package.json
@@ -20,5 +20,7 @@
   },
   "devDependencies": {
     "prisma": "^6.10.1"
-  }
+  },
+  "type": "commonjs"
+
 }
